+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog][Keep a Changelog] and this project adh
 - Unit tests for GetTag and GetImage
 
 ### Changed
+- Return TagResponseModel from GetTag instead of TagModel
 - Restructure project and namespaces
 - Configure build stages